Rename to GenericContainerArray.h
